libs/dingdoor-utils-package/pyproject.toml
[project]
name = "dingdoor-utils-package"
version = "0.1.0"
description = "Package helpers for Dingdoor"
readme = "README.md"
authors = [
    { name = "Paulo Costa", email = "pscostam11@gmail.com" }
]
requires-python = ">=3.11"
dependencies = ["google-cloud-bigquery>=3.22.0"]

[project.scripts]
dingdoor-utils-package = "dingdoor_utils_package:main"

[build-system]
requires = ["uv_build>=0.9.0,<0.10.0"]
build-backend = "uv_build"
